10:21

+++ app/src/main/java/com/htyhbz/yhyg/activity/order/OrderSettlementActivity.java
@@ -87,7 +87,7 @@ public class OrderSettlementActivity extends BaseActivity{
                 Product product=new Product();
                 product.setproductId(obj.getInt("productId"));
                 product.setproductName(obj.getString("productName"));
-                product.setproductPrice(obj.getDouble("productPrice"));
+                product.setproductPrice(obj.getInt("productPrice"));
                 product.setproductPictureUrl(obj.getString("productPictureUrl"));
                 product.setorderProductCount(obj.getInt("shoppingsingleTotal"));
                 orderProductionsID.append(obj.getInt("productId") + ",");
